chore: use controller-runtime to define chaos
